qt4: Upgraded qt to 4.7.1, rewrote recipes, seen QT4E work on arm arch.
Note: qt4-x11-free does NOT work!. it is included to further development
and further work on it will continue when x11 works.

Most importantly; the qt4E recipe (including qt4.inc) is written to support
one specific set of requirements (make qt quick 4.7.2 work with as little
overhead as possible). Lots of work still exists to make it possible to
build everything and then afterwards select which specific components one
wants.

The problem atm, is that the times it takes to build a full qt with
ALL modules simply just takes too long when changing the recipe (on my
laptor between 5.5 and 6 hours for just qt). Reconsider solutions
to splitting the qt recipe into many smaller recipes to make it possible to
build e.g. only webkit hereby avoiding building that which is not wanted.